Rewrite folder structure
