+++ tests/shop/advantage/fixtures/contract-item-legacy.json
