+++ tests/shop/advantage/fixtures/contract-item-shop.json
