+++ tests/shop/advantage/fixtures/contract-item-trial.json
