+++ tests/shop/advantage/fixtures/contract.json
