+++ tests/shop/advantage/fixtures/contracts.json
