+++ tests/shop/advantage/fixtures/offer.json
