+++ tests/shop/advantage/fixtures/offers.json
