+++ tests/shop/advantage/fixtures/product.json
