+++ tests/shop/advantage/fixtures/purchase.json
