+++ tests/shop/advantage/fixtures/purchases.json
