+++ tests/shop/advantage/fixtures/subscription-auto-renewal.json
