+++ tests/shop/cube/test_cube.py
